+++ src/directives/formly-form.js
@@ -1,6 +1,6 @@
 'use strict';
 angular.module('formly.render')
-.directive('formlyForm', function formlyForm() {
+.directive('formlyForm', function formlyForm($parse) {
 	return {
 		restrict: 'AE',
 		templateUrl: 'directives/formly-form.html',
@@ -17,6 +17,16 @@ angular.module('formly.render')
 					//Post gets called after angular has created the FormController
 					//Now pass the FormController back up to the parent scope
 					scope.formOnParentScope = scope[attr.name];
+					angular.forEach(scope.fields, function(field) {
+						if (field.hideExpression) {
+						var getter = $parse(field.hideExpression);
+						scope.$watch(function() {
+							return getter(scope.result);
+						}, function(hide) {
+							field.hide = hide;
+						});
+						}
+					});
 				}
 			}
 		}