+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Version numbers correspond to `bower.json` version
 
 ## Features
 
-Added `hide` property on fields. Allows devs to conditionally show fields.
+Added `hide` property on fields. Allows devs to conditionally show fields. Also adding `hideExpression` property which is an expression that will be evaluated on the result object (using $parse) to make the api simpler to use.
 
 ## Bug Fixes