+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 Version numbers correspond to `bower.json` version
 
+# 0.0.10
+
+## Features
+
+Added `hide` property on fields. Allows devs to conditionally show fields. Also adding `hideExpression` property which is an expression that will be evaluated on the result object (using $parse) to make the api simpler to use.
+
+## Bug Fixes
+
+## Breaking Changes
+
 # 0.0.1
 
 ## Features